refactor(writer): Create submodule
Extract functions for writing output to their own submodule.
No tests for these functions as of yet..
